frameworks/base: TRDS 3.0 and new theme engine preparation (1/3)
For the one who do not know this feature see as reference
the first two versions here:

SlimRoms/frameworks_base@b330fde

SlimRoms/frameworks_base@3194015

we got now over a year now a lot experience how to handle the qualifier
fallback
as well how to theme it and how to handle other apps. Based on this
knowledge we
redid some parts of TRDS engine with following new features/additions
and fixes
to have now a final perfect working TRDS

- compared to the old versions where it was needed on some apps to
restart them
in background this is not needed any more. All is handled about the
internal
AOSP resource handling and on qualifier change the caches are cleared
and the apps get an onconfigurationchanged to restart the activity
like it is doing on orientation change eg
- inverted folder is now depreciated and the new folders are holodark
and hololight which allows on the one hand a dark theme (at least
how TRDS started) and a light based theme hololight
- This name change is mainly because of the new theme engine which
will come which uses this as base to add 2 external themes at the same
time and give the user the ability to switch between them. But this is
still in early stages and this commit is just the preparation for it
- Two new automodes are now integrated. User can enable the twilight
modus
which enables/disables holodark/hololight depending on the
daytime...means
on the evening it switches to holodark on normal day it is hololight.
The other automode is based on light conditions....means that if it is
dark outside it switches to holodark....if we have bright light
conditions
it switches immediatelly to hololight. Due of the resource cache
handling now
this happens immediatelly for the user and feels completly system
integrated
and happens without delay. A bit comparable to some navigation tool for
cars
which can switch automatically from dark to light and back
- currently the auto modes as well the manual mode switches between
holodark
and hololight. Due that we do not have currently anything hololight
themed
in the source hololight = stock theme. We will give the user the ability
later when done to choose between which modes the user want to switch
(stock/holodark/hololight). But this does not make sense at the moment.
As soon as we have done the light theme or theme engine (lets see what
comes first) we will open the ability to choose it. This is only a small
code change
- a lot small tweaks and fixes went as well into the new version

Whatever this commit as usual adds the support for it. The themeing is
done in the
other commits by kroz which is as well way more extended then on the
older versions

cheers

PatchSets:
- minor code cleanup

PatchSet5:
- use one value for auto modes

PatchSet6:
- kill resource caches on configuration change

PatchSet7:
- register settings observer lol

PatchSet8:
- fix not assigment of config mode change (typo lol)

PatchsSet 9 -11:
- playing around aosp qualifier is not that flexible then I thought
whatever back in most parts to original implementation handling

PatchSet 12:
- now no restarts of any application needed any more like it was on
4.3...yeaaah
- fix wrong shifting in native adapter
- fix wrong call of twilightmode
- disable debug
- cleanup

PatchSet 13:
- fix lightsensor

Patchset 14:
- no need currently to recreate the navbar we can do if a theme needs it

Patchset 16:
- synchronize on sending config change
- only send if system is ready
- fire first update when system is ready

-rebase