fb: fix and pimp custom lockscreen handle (1/2)
problem on current implementation is that it is not
multiDpi compatible. As well it looks a bit weird if
the user set a square image always. This commit does

- introduce central circleimage method to get circle image crops
- before it was assumed that the height of the handle image is
108px (mdpi base) which is correct but the original handle
has a lot unused space around to fit better into the view.
Whatever after messurment of the real height 68px is correct
of course calculated against the current density
- both above combined results into a always nice looking
custom circle icon on the lockscreen dpi and device undependent
- saved file in settings calculates the needed image size
base on the current density.

PS1: Initial State